+++ packages/mobx/src/types/observablearray.ts
@@ -22,13 +22,12 @@ import {
     registerListener,
     spyReportEnd,
     spyReportStart,
-    allowStateChangesStart,
-    allowStateChangesEnd,
     assertProxies,
     reserveArrayBuffer,
     hasProp,
     die,
-    globalState
+    globalState,
+    initObservable
 } from "../internal"
 
 const SPLICE = "splice"
@@ -406,16 +405,16 @@ export function createObservableArray<T>(
     owned = false
 ): IObservableArray<T> {
     assertProxies()
-    const adm = new ObservableArrayAdministration(name, enhancer, owned, false)
-    addHiddenFinalProp(adm.values_, $mobx, adm)
-    const proxy = new Proxy(adm.values_, arrayTraps) as any
-    adm.proxy_ = proxy
-    if (initialValues && initialValues.length) {
-        const prev = allowStateChangesStart(true)
-        adm.spliceWithArray_(0, 0, initialValues)
-        allowStateChangesEnd(prev)
-    }
-    return proxy
+    return initObservable(() => {
+        const adm = new ObservableArrayAdministration(name, enhancer, owned, false)
+        addHiddenFinalProp(adm.values_, $mobx, adm)
+        const proxy = new Proxy(adm.values_, arrayTraps) as any
+        adm.proxy_ = proxy
+        if (initialValues && initialValues.length) {
+            adm.spliceWithArray_(0, 0, initialValues)
+        }
+        return proxy
+    })
 }
 
 // eslint-disable-next-line